lhapdf6if.f and lhapdf6ifcc.cc: modified to implement the option to extract alphas from LHAPDF also for version lhapdf v6.
setstrongcoupl.f: modified to allow the use of alphas from lhapdf (by adding in the input card 'alphas_from_lhapdf 1'). This is an "experimental" option, that might be useful/needed for some processes, but by default the code runs exactly as it used to do. One needs to be careful when using this option, as the behaviour of alphaspdf might be slightly different from the normal powheg running, especially at small scales. This might have an impact on the generation of the hardest radiation (for instance producing side-effects with the upper-bounds for generating the hardest radiation).



git-svn-id: svn://powhegbox.mib.infn.it/trunk/POWHEG-BOX-V2@3427 9e129b38-c2c0-4eec-9301-dc54fda404de