A process-independent Herwig7.1 interface.
In principle, it is possible to use the standard LesHouches library provided by
ThePEG (see LHE.in in the Herwig7 Contrib directory) to read in .lhe files
generated by POWHEG BOX and to shower them in Herwig. This is not very convenient,
however, due to the following:
a.) The analysis needs to be re-implemented in native Herwig, or other tools
suitable for analyses like Rivet.
b.) The LHE multi-weight standard employed by POWHEG is not compatible with the
one employed by ThePEG. 
c.) The ThePEG library also lacks certain other features (not explained here)
like ub_corr_* factors. 
d.) If Herwig fails to shower the event, it discards it by default, which may
not be desirable. 

This interface solves all of the above and more.



git-svn-id: svn://powhegbox.mib.infn.it/trunk/POWHEG-BOX-V2@3591 9e129b38-c2c0-4eec-9301-dc54fda404de