Few improvements added: 1) If a file named include/ProcessCitation.f is present in the process directory, it is included by the main program as first instruction. You can use this to print a banner for the process
2) A mechanism to discard outliers when using storemintupb has been
setup. In order to use it include
StoredubNoOutliers 1
in the powheg.input.
It is off by default. If outliers are found, their number is printed
in the st3 counters, and their value is printed to standard output.

Other minor changes: document on bound violation fixes and on Compression and Weights
have been improved. A comment in btilde.f has been fixed.



git-svn-id: svn://powhegbox.mib.infn.it/trunk/POWHEG-BOX-V2@3681 9e129b38-c2c0-4eec-9301-dc54fda404de